Refactoring concurrent_009 into base rule multiline_conditional_alignment.
