+++ vsg/rules/__init__.py
@@ -69,6 +69,7 @@ from .multiline_structure import multiline_structure
 from .multiline_simple_structure import multiline_simple_structure
 from .multiline_constraint_structure import multiline_constraint_structure
 from .multiline_procedure_call_structure import multiline_procedure_call_structure
+from .multiline_conditional_alignment import multiline_conditional_alignment
 from .remove_carriage_return_after_token import remove_carriage_return_after_token
 from .move_token_to_the_right_of_several_possible_tokens_if_it_exists_between_tokens import move_token_to_the_right_of_several_possible_tokens_if_it_exists_between_tokens
 from .number_of_lines_between_tokens import number_of_lines_between_tokens
